packages/ocdsextensionregistry/ocdsextensionregistry-0.1.3.tar.gz/ocdsextensionregistry-0.1.3/ocdsextensionregistry/extension.py
class Extension:
    def __init__(self, data):
        """
        Accepts a row from extensions.csv and assigns values to properties.
        """
        self.id = data['Id']
        self.category = data['Category']
        self.core = data['Core'] == 'true'

    def __repr__(self):
        return self.id

    def as_dict(self):
        """
        Returns the object's properties as a dictionary.

        This method is defined to match the method in `ExtensionVersion`.
        """
        return self.__dict__
